+++ tests/test_cfg_global.py
@@ -18,7 +18,7 @@ from logging import DEBUG
 from logging import debug
 from tempfile import TemporaryDirectory
 from timetracker.cfg.cfg_global import CfgGlobal
-from timetracker.cfg.cfg import CfgProj
+from timetracker.cfg.cfg_local import CfgProj
 from timetracker.cfg.utils import get_relpath_adj
 #from timetracker.cli import Cli
 #from timetracker.filemgr import FileMgr