+++ tests/test_usecases.py
@@ -7,10 +7,10 @@ from os import environ
 from collections import namedtuple
 from datetime import timedelta
 from timeit import default_timer
-from timetracker.cfg.cfg import CfgProj
+from timetracker.cfg.cfg_global import CfgGlobal
+from timetracker.cfg.cfg_local import CfgProj
 from timetracker.cli import Cli
 from timetracker.filemgr import FileMgr
-from timetracker.cfg.cfg_global import CfgGlobal
 
 # pylint: disable=fixme