+++ timetracker/trk.py
@@ -5,11 +5,11 @@ __author__ = "DV Klopfenstein, PhD"
 
 from os.path import exists
 from logging import error
-##from logging import DEBUG
-##from logging import basicConfig
+#from logging import DEBUG
+#from logging import basicConfig
 from timetracker.filemgr import FileMgr
-from timetracker.cfg.cfg import CfgProj
 from timetracker.cfg.cfg_global import CfgGlobal
+from timetracker.cfg.cfg_local import CfgProj
 from timetracker.cli import Cli
 from timetracker.cmd.init import run_init
 from timetracker.cmd.start import run_start